Fix type error
In handleBlockUser, handleDeleteChat is now only called if both chatToDelete & its _id are defined

+++ src/Components/Pages/OtherUserProfile/OtherUserProfile.tsx
@@ -597,7 +597,7 @@ const OtherUserProfile = () => {
         )[0]
       : undefined;
 
-    if (chatToDelete) {
+    if (chatToDelete && chatToDelete._id) {
       handleDeleteChat(chatToDelete._id.toString());
     }
